Move action_fallback macro to public api
